+++ src/meta/service/src/meta_service/mod.rs
@@ -13,8 +13,8 @@
 // limitations under the License.
 
 pub use forwarder::MetaForwarder;
-pub use meta_service_impl::RaftServiceImpl;
-pub use raftmeta::MetaNode;
+pub use meta_node::MetaNode;
+pub use raft_service_impl::RaftServiceImpl;
 
 pub use crate::message::ForwardRequest;
 pub use crate::message::ForwardRequestBody;
@@ -24,6 +24,6 @@ pub use crate::message::LeaveRequest;
 mod errors;
 mod forwarder;
 pub mod meta_leader;
+pub mod meta_node;
 mod meta_node_kv_api_impl;
-pub mod meta_service_impl;
-pub mod raftmeta;
+pub mod raft_service_impl;
