+++ src/meta/service/tests/it/store.rs
@@ -29,8 +29,8 @@ use common_meta_types::StoredMembership;
 use common_meta_types::TypeConfig;
 use common_meta_types::Vote;
 use common_tracing::func_name;
-use databend_meta::meta_service::raftmeta::LogStore;
-use databend_meta::meta_service::raftmeta::SMStore;
+use databend_meta::meta_service::meta_node::LogStore;
+use databend_meta::meta_service::meta_node::SMStore;
 use databend_meta::store::RaftStore;
 use databend_meta::Opened;
 use log::debug;
